api: SFTP Port int→int32 and add FileMode/UID/GID validation
M1: Change SFTPOptions.Port from int to int32 for consistency with
other port fields and Kubernetes API conventions.
M2: Add kubebuilder validation markers for FileMode (0-511),
DirectoryMode (0-511), FileUID (>=0), and FileGID (>=0).
Regenerate CRDs.

+++ internal/kopia/server/manager.go
@@ -460,7 +460,7 @@ func (m *KopiaServerManager) writeFilesystemServerScript(b *strings.Builder, rep
 // writeSFTPServerScript writes the SFTP-mode repository connect/create script.
 func (m *KopiaServerManager) writeSFTPServerScript(b *strings.Builder, repo *backupv1alpha1.KopiaRepository) {
 	cacheFlags := BuildCacheFlags(repo.Spec.Caching)
-	port := 22
+	port := int32(22)
 	if repo.Spec.SFTPOptions.Port > 0 {
 		port = repo.Spec.SFTPOptions.Port
 	}
